requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
deep-sort-realtime==1.3.2
distro==1.9.0
et_xmlfile==2.0.0
filelock==3.18.0
Flask==3.1.1
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
fonttools==4.59.0
fpdf==1.7.2
fsspec==2025.3.2
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
kiwisolver==1.4.8
lxml==5.3.2
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.5
openai==1.76.2
opencv-python==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pillow==11.1.0
psutil==7.0.0
py-cpuinfo==9.0.0
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-engineio==4.12.2
python-socketio==5.13.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
reportlab==4.3.1
requests==2.32.3
safetensors==0.5.3
scipy==1.16.0
setuptools==80.1.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sympy==1.14.0
tokenizers==0.21.1
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.51.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.168
ultralytics-thop==2.0.14
urllib3==2.4.0
Werkzeug==3.1.3
wsproto==1.2.0
